[dagster-airlift] move defs/mapping construction out of AirflowDefinitionsData
